Move keysend tests to payment_tests.rs
